Fix bad interface function type in netseqcheck

+++ src/main/java/org/gr1m/mc/mup/tweaks/netseqcheck/INetSequenceHandler.java
@@ -6,5 +6,5 @@ public interface INetSequenceHandler
     void incrSendingSequenceNumber();
 
     int getCheckingSequenceNumber();
-    void setCheckingSequenceNumber(long sequenceIn);
+    void setCheckingSequenceNumber(int sequenceIn);
 }
